rebased vertex code
moved from inside the EWEModel class to a new EWE_Vertex class, much less bloat in model. No longer have to scroll 400 lines just to see model code
