+++ src/EWE_Model.cpp
@@ -20,24 +20,24 @@
 
 namespace std {
     template <>
-    struct hash<EWE::EWEModel::Vertex> {
-        size_t operator()(EWE::EWEModel::Vertex const& vertex) const {
+    struct hash<EWE::Vertex> {
+        size_t operator()(EWE::Vertex const& vertex) const {
             size_t seed = 0;
             EWE::hashCombine(seed, vertex.position, vertex.color, vertex.normal, vertex.uv);
             return seed;
         }
     };
     template<>
-    struct hash<EWE::EWEModel::simpleVertex> {
-        size_t operator()(EWE::EWEModel::simpleVertex const& vertex) const {
+    struct hash<EWE::simpleVertex> {
+        size_t operator()(EWE::simpleVertex const& vertex) const {
             size_t seed = 0;
             EWE::hashCombine(seed, vertex.position);
             return seed;
         }
     };
     template<>
-    struct hash<EWE::EWEModel::GrassVertex> {
-        size_t operator()(EWE::EWEModel::GrassVertex const& vertex) const {
+    struct hash<EWE::GrassVertex> {
+        size_t operator()(EWE::GrassVertex const& vertex) const {
             size_t seed = 0;
             EWE::hashCombine(seed, vertex.position, vertex.color);
             return seed;
@@ -72,7 +72,7 @@ namespace EWE {
 
         glm::vec3 color = { 1.f,1.f,1.f };
         glm::vec3 normal = { 0.f, -1.f, 0.f };
-        std::vector<EWEModel::Vertex> vertices;
+        std::vector<Vertex> vertices;
         vertices.push_back({ { 0.0f,0.0f,0.0f }, normal, { 0.5f,0.5f }, color });
 
         float angle = glm::two_pi<float>() / points;
@@ -289,171 +289,6 @@ namespace EWE {
         else {
             vkCmdDraw(commandBuffer, vertexCount, instanceCount, 0, 0);
         }
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::simpleVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(simpleVertex, position) });
-
-        return attributeDescriptions;
-    }
-
-
-    std::vector<VkVertexInputBindingDescription> EWEModel::GrassVertex::getBindingDescriptions() { //still here because instanced
-        std::vector<VkVertexInputBindingDescription> bindingDescriptions(2);
-        bindingDescriptions[0].binding = 0;
-        bindingDescriptions[0].stride = sizeof(GrassVertex);
-        bindingDescriptions[0].inputRate = VK_VERTEX_INPUT_RATE_VERTEX; 
-        
-        bindingDescriptions[1].binding = 1;
-        bindingDescriptions[1].stride = sizeof(GrassInstance);
-        bindingDescriptions[1].inputRate = VK_VERTEX_INPUT_RATE_INSTANCE;
-        
-        return bindingDescriptions;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::GrassVertex::getAttributeDescriptions() {
-
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions = {
-            { 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(GrassVertex, position) },
-            { 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(GrassVertex, color) },
-            //{ 1, 0, VK_FORMAT_R32_SFLOAT, sizeof(glm::vec3) * 3 },
-
-            //instance
-            { 2, 1, VK_FORMAT_R32G32B32A32_SFLOAT, 0},
-            { 3, 1, VK_FORMAT_R32G32B32A32_SFLOAT, sizeof(glm::vec4)},
-            { 4, 1, VK_FORMAT_R32G32B32A32_SFLOAT, sizeof(glm::vec4) * 2},
-            { 5, 1, VK_FORMAT_R32G32B32A32_SFLOAT, sizeof(glm::vec4) * 3},
-            {6, 1, VK_FORMAT_R32G32_SFLOAT, sizeof(glm::vec4) * 4}
-        };
-        return attributeDescriptions;
-    }
-    std::vector<VkVertexInputBindingDescription> EWEModel::LeafVertex::getBindingDescriptions() { //still here because instanced
-        std::vector<VkVertexInputBindingDescription> bindingDescriptions(1);
-        bindingDescriptions[0].binding = 0;
-        bindingDescriptions[0].stride = sizeof(LeafVertex);
-        bindingDescriptions[0].inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-        /*
-        bindingDescriptions[1].binding = 1;
-        bindingDescriptions[1].stride = sizeof(LeafInstance);
-        bindingDescriptions[1].inputRate = VK_VERTEX_INPUT_RATE_INSTANCE;
-        */
-        return bindingDescriptions;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::LeafVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions = {
-            { 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(LeafVertex, position) },
-            { 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(LeafVertex, normal) },
-            { 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(LeafVertex, uv) },
-
-            //instance
-            /*
-            { 3, 1, VK_FORMAT_R32G32B32A32_SFLOAT, 0},
-            { 4, 1, VK_FORMAT_R32G32B32A32_SFLOAT, sizeof(glm::vec4)},
-            { 5, 1, VK_FORMAT_R32G32B32A32_SFLOAT, sizeof(glm::vec4) * 2},
-            { 6, 1, VK_FORMAT_R32G32B32A32_SFLOAT, sizeof(glm::vec4) * 3},
-            */
-        };
-
-        return attributeDescriptions;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::EffectVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(EffectVertex, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(EffectVertex, uv) });
-
-        return attributeDescriptions;
-    }
-
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::Vertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, normal) });
-        attributeDescriptions.push_back({ 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(Vertex, uv) });
-        attributeDescriptions.push_back({ 3, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(Vertex, color) });
-
-        return attributeDescriptions;
-    }
-    std::vector<VkVertexInputAttributeDescription> EWEModel::skyVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(skyVertex, position) });
-
-        return attributeDescriptions;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::boneVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(boneVertex, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(boneVertex, normal) });
-        attributeDescriptions.push_back({ 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(boneVertex, uv) });
-        attributeDescriptions.push_back({ 3, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(boneVertex, tangent) });
-        attributeDescriptions.push_back({ 4, 0, VK_FORMAT_R32G32B32A32_SINT, offsetof(boneVertex, m_BoneIDs) });
-        attributeDescriptions.push_back({ 5, 0, VK_FORMAT_R32G32B32A32_SFLOAT, offsetof(boneVertex, m_Weights) });
-
-        return attributeDescriptions;
-    }
-    std::vector<VkVertexInputAttributeDescription> EWEModel::glmVertexNoTangent::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(glmVertexNoTangent, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(glmVertexNoTangent, normal) });
-        attributeDescriptions.push_back({ 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(glmVertexNoTangent, uv) });
-        attributeDescriptions.push_back({ 3, 0, VK_FORMAT_R32G32B32A32_SINT, offsetof(glmVertexNoTangent, m_BoneIDs) });
-        attributeDescriptions.push_back({ 4, 0, VK_FORMAT_R32G32B32A32_SFLOAT, offsetof(glmVertexNoTangent, m_Weights) });
-
-        return attributeDescriptions;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::AVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(AVertex, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(AVertex, normal) });
-        attributeDescriptions.push_back({ 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(AVertex, uv) });
-        attributeDescriptions.push_back({ 3, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(AVertex, tangent) });
-
-        return attributeDescriptions;
-    }
-    std::vector<VkVertexInputAttributeDescription> EWEModel::AVertexNT::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(AVertexNT, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(AVertexNT, normal) });
-        attributeDescriptions.push_back({ 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(AVertexNT, uv) });
-
-        return attributeDescriptions;
-    }
-    std::vector<VkVertexInputAttributeDescription> bobVertex::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions{};
-
-        attributeDescriptions.push_back({ 0, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(bobVertex, position) });
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(bobVertex, normal) });
-        attributeDescriptions.push_back({ 2, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(bobVertex, uv) });
-        attributeDescriptions.push_back({ 3, 0, VK_FORMAT_R32G32B32_SFLOAT, offsetof(bobVertex, tangent) });
-        attributeDescriptions.push_back({ 4, 0, VK_FORMAT_R32G32B32A32_SINT, offsetof(bobVertex, m_BoneIDs) });
-        attributeDescriptions.push_back({ 5, 0, VK_FORMAT_R32G32B32A32_SFLOAT, offsetof(bobVertex, m_Weights) });
-
-        return attributeDescriptions;
-    }
-
-    std::vector<VkVertexInputAttributeDescription> EWEModel::VertexUI::getAttributeDescriptions() {
-        std::vector<VkVertexInputAttributeDescription> attributeDescriptions(1);
-        attributeDescriptions[0].binding = 0;
-        attributeDescriptions[0].location = 0;
-        attributeDescriptions[0].format = VK_FORMAT_R32G32_SFLOAT;
-        attributeDescriptions[0].offset = offsetof(VertexUI, position);
-
-        attributeDescriptions.push_back({ 1, 0, VK_FORMAT_R32G32_SFLOAT, offsetof(VertexUI, uv) });
-
-        return attributeDescriptions;
     }
 
     void EWEModel::Builder::loadModel(const std::string& filepath) {
