+++ src/StaticRenderingSystem.cpp
@@ -1,0 +1,2 @@
+#include "EWEngine/systems/StaticRendering/StaticRenderingSystem.h"
+
